Remove dependencies' cycles
