+++ src/animations/spring.js
@@ -18,7 +18,7 @@ import {
   greaterThan,
 } from '../base';
 import { min, abs } from '../derived';
-import AnimatedValue from '../core/AnimatedValue';
+import AnimatedValue from '../core/InternalAnimatedValue';
 
 const MAX_STEPS_MS = 64;
