+++ src/core/AnimatedBlock.js
@@ -1,5 +1,6 @@
 import AnimatedNode from './AnimatedNode';
-import { val } from '../utils';
+import { val } from '../val';
+import InternalAnimatedValue from './InternalAnimatedValue';
 
 export default class AnimatedBlock extends AnimatedNode {
   _array;
@@ -16,4 +17,25 @@ export default class AnimatedBlock extends AnimatedNode {
     });
     return result;
   }
+}
+
+export function createAnimatedBlock(items) {
+  return adapt(items);
+}
+
+function nodify(v) {
+  if (typeof v === 'object' && v.__isProxy) {
+    if (!v.__val) {
+      v.__val = new InternalAnimatedValue(0);
+    }
+    return v.__val;
+  }
+  // TODO: cache some typical static values (e.g. 0, 1, -1)
+  return v instanceof AnimatedNode ? v : new InternalAnimatedValue(v);
+}
+
+export function adapt(v) {
+  return Array.isArray(v)
+    ? new AnimatedBlock(v.map(node => adapt(node)))
+    : nodify(v);
 }
