+++ src/core/AnimatedCall.js
@@ -1,5 +1,5 @@
 import ReanimatedEventEmitter from '../ReanimatedEventEmitter';
-import { val } from '../utils';
+import { val } from '../val';
 import AnimatedNode from './AnimatedNode';
 
 const NODE_MAPPING = new Map();
@@ -39,4 +39,8 @@ export default class AnimatedCall extends AnimatedNode {
     this._callback(this._args.map(val));
     return 0;
   }
+}
+
+export function createAnimatedCall(args, func) {
+  return new AnimatedCall(args, func);
 }
