+++ src/core/AnimatedConcat.js
@@ -1,7 +1,12 @@
 import AnimatedNode from './AnimatedNode';
+import { adapt } from '../core/AnimatedBlock';
 
 export default class AnimatedConcat extends AnimatedNode {
   constructor(input) {
     super({ type: 'concat', input: input.map(n => n.__nodeID) }, input);
   }
+}
+
+export function createAnimatedConcat(...args) {
+  return new AnimatedConcat(args.map(adapt));
 }
