+++ src/core/AnimatedDebug.js
@@ -1,5 +1,8 @@
-import { val } from '../utils';
+import { val } from '../val';
 import AnimatedNode from './AnimatedNode';
+import { adapt } from '../core/AnimatedBlock';
+import { createAnimatedBlock as block } from './AnimatedBlock';
+import { createAnimatedCall as call } from './AnimatedCall';
 
 export default class AnimatedDebug extends AnimatedNode {
   _message;
@@ -16,4 +19,28 @@ export default class AnimatedDebug extends AnimatedNode {
     console.log(this._message, value);
     return value;
   }
+}
+
+export function createAnimatedDebug(message, value) {
+  if (__DEV__) {
+    const runningInRemoteDebugger = typeof atob !== 'undefined';
+    // hack to detect if app is running in remote debugger
+    // https://stackoverflow.com/questions/39022216
+
+    const runningInExpoShell =
+      global.Expo && global.Expo.Constants.appOwnership !== 'standalone';
+
+    if (runningInRemoteDebugger || runningInExpoShell) {
+      // When running in expo or remote debugger we use JS console.log to output variables
+      // otherwise we output to the native console using native debug node
+      return block([
+        call([value], ([a]) => console.log(`${message} ${a}`)),
+        value,
+      ]);
+    } else {
+      return new AnimatedDebug(message, adapt(value));
+    }
+  }
+  // Debugging is disabled in PROD
+  return value;
 }
