+++ src/core/AnimatedSet.js
@@ -1,5 +1,6 @@
 import AnimatedNode from './AnimatedNode';
-import { val } from '../utils';
+import { val } from '../val';
+import { adapt } from '../core/AnimatedBlock';
 
 export default class AnimatedSet extends AnimatedNode {
   _what;
@@ -16,4 +17,8 @@ export default class AnimatedSet extends AnimatedNode {
     this._what._updateValue(newValue);
     return newValue;
   }
+}
+
+export function createAnimatedSet(what, value) {
+  return new AnimatedSet(what, adapt(value));
 }
