+++ src/core/AnimatedValue.js
@@ -1,56 +1,9 @@
-import AnimatedNode from './AnimatedNode';
-import { set } from '../base';
-import { val } from '../utils';
-import { evaluateOnce } from '../derived/evaluateOnce';
+import { createAnimatedSet as set } from '../core/AnimatedSet';
 import interpolate from '../derived/interpolate';
-import ReanimatedModule from '../ReanimatedModule';
+import InternalAnimatedValue from './InternalAnimatedValue';
+import { evaluateOnce } from '../derived/evaluateOnce';
 
-function sanitizeValue(value) {
-  return value === null || value === undefined || typeof value === 'string'
-    ? value
-    : Number(value);
-}
-
-export default class AnimatedValue extends AnimatedNode {
-  constructor(value) {
-    super({ type: 'value', value: sanitizeValue(value) });
-    this._startingValue = this._value = value;
-    this._animation = null;
-  }
-
-  __detach() {
-    ReanimatedModule.getValue(
-      this.__nodeID,
-      val => (this.__nodeConfig.value = val)
-    );
-    this.__detachAnimation(this._animation);
-    super.__detach();
-  }
-
-  __detachAnimation(animation) {
-    animation && animation.__detach();
-    if (this._animation === animation) {
-      this._animation = null;
-    }
-  }
-
-  __attachAnimation(animation) {
-    this.__detachAnimation(this._animation);
-    this._animation = animation;
-  }
-
-  __onEvaluate() {
-    if (this.__inputNodes && this.__inputNodes.length) {
-      this.__inputNodes.forEach(val);
-    }
-    return this._value + this._offset;
-  }
-
-  _updateValue(value) {
-    this._value = value;
-    this.__forceUpdateCache(value);
-  }
-
+export default class AnimatedValue extends InternalAnimatedValue {
   setValue(value) {
     this.__detachAnimation(this._animation);
     evaluateOnce(set(this, value), this);
