+++ src/derived/acc.js
@@ -1,5 +1,5 @@
 import { set, add } from '../base';
-import AnimatedValue from '../core/AnimatedValue';
+import AnimatedValue from '../core/InternalAnimatedValue';
 
 export default function acc(v) {
   const acc = new AnimatedValue(0);
