+++ src/derived/onChange.js
@@ -1,5 +1,5 @@
 import { block, cond, defined, neq, not, set } from '../base';
-import AnimatedValue from '../core/AnimatedValue';
+import AnimatedValue from '../core/InternalAnimatedValue';
 
 export default function onChange(value, action) {
   const prevValue = new AnimatedValue();
